+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Changed
 - allow bmp and webp for backer file upload
+- period contract overview
+  - show green contract icon even without timestamp
+  - show blue ad data icon if files are available
 
 ## v0.6.3 - 2024-03-09
 ### Changed